pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "orbit"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.0",
    "google-adk>=0.1.0",
    "redis>=5.0.0",
    "asyncpg>=0.29.0",
    "pgvector>=0.3.0",
    "psycopg2-binary>=2.9.9",
    "boto3>=1.34.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.29.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py311"
